move files around, refactor scss files and add tsconfig
