+++ src/App.js
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router-dom';
 import NavigationBar from './components/NavigationBar';
 import Routes from './components/Routes';
 import 'antd/dist/antd.css';
-import './App.scss';
+import './css/App.scss';
 
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { fab } from '@fortawesome/free-brands-svg-icons'
