--- src/components/Contact.scss
@@ -1,75 +1,0 @@
-@import "../utilities/themes";
-
-.contact-container {
-    padding-top: 7em;
-    p {
-        @include themed() {
-            color: t($nav-text);
-        }
-    }
-    a {
-        @include themed() {
-            color: t($nav-text);
-        }
-        text-decoration: underline;
-        &:hover {
-            text-decoration: underline;
-            @include themed() {
-                color: t($link-hover);
-            }
-        }
-    }
-}
-
-.form-wrapper {
-    height: auto;
-    width: 100%;
-    .form-group {
-        align-content: center;
-    }
-    .form-control {
-        width: 50%;
-        margin: 0 auto;
-    }
-    .message-box {
-        height: 15em !important;
-    }
-    .btn-secondary {
-        width: 10em;
-        height: 3em;
-        @include themed() {
-            background-color: t($button-background);
-            border-color: t($button-background);
-            color: t($button-text);
-        }
-        &:hover {
-            @include themed() {
-                background-color: t($button-background-hover);
-                border-color: t($button-background-hover);
-                color: t($button-text-hover);
-            }
-        }
-    }
-}
-
-@media only screen and (max-width: 768px) {
-    .contact-container {
-        p {
-            padding: 0.5em;
-        }
-        a {
-            @include themed() {
-                color: t($link-hover);
-            }
-        }
-    }
-    .form-wrapper {
-        .form-control {
-            width: 80%;
-            margin: 0 auto;
-        }
-        .message-box {
-            height: 10em !important;
-        }
-    }
-}
