--- src/components/Home.scss
@@ -1,21 +1,0 @@
-.home-container {
-    padding-top: 6em;
-    img {
-        height: 25em;
-        border-radius: 50%;
-        padding: 1em;
-    }
-    p {
-        text-align: left;
-    }
-}
-
-@media only screen and (max-width: 768px) {
-    .home-container {
-        img {
-            height: 20em;
-            margin: -2em 0em -5em 0em;
-            padding: 0em;
-        }
-    }
-}
