--- src/components/Photography.scss
@@ -1,45 +1,0 @@
-@import "../utilities/themes";
-
-.photography-container {
-    padding-top: 7em;
-}
-
-.gallery {
-    max-width: 90%;
-    margin: auto;
-    img {
-        @include themed() {
-            background-color: t($photography-frames);
-        }  
-        width: 28em;
-        border-radius: 0.5em;
-        padding: 1em 1em 1em 1em;
-        margin: 0.5em 0.5em 2em 0.5em;
-    }
-}
-
-@media only screen and (max-width: 768px) {
-    .gallery {
-        img {
-            max-width: 90%;
-        }
-    }
-}
-
-@media only screen and (min-width: 768px) and (max-width: 1024px) {
-  .gallery {
-        img {
-            max-width: 90%;
-        }
-    }
-}
-
-@media only screen and (min-width: 1024px) and (max-width: 1400px) {
-    .gallery {
-          img {
-              max-width: 85%;
-          }
-      }
-  }
-
-
