--- src/components/Resume.scss
@@ -1,45 +1,0 @@
-@import "../utilities/themes";
-
-.resume-container {
-    padding-top: 7em;
-    h1 {
-        margin-bottom: 0em;
-    }
-    img {
-        height: 60em;
-        padding-bottom: 3em;
-    }
-    a {
-        @include themed() {
-            color: t($main-paragraph);
-        }
-        text-decoration: underline;
-        &:hover {
-            @include themed() {
-                color: t($link-hover);
-            }
-            text-decoration: underline;
-        }
-    }
-    p {
-        @include themed() {
-            color: t($main-paragraph);
-        }
-    }
-}
-
-@media only screen and (max-width: 768px) {
-    .resume-container {
-        img {
-            height: 40em;
-        }
-        a {
-            @include themed() {
-                color: t($link-hover);
-            }
-        }
-        p {
-            padding: 0em 1em 0em 1em;
-        }
-    }
-}
