+++ src/css/App.scss
@@ -1,5 +1,5 @@
 @import '~antd/dist/antd.css';
-@import './utilities/themes';
+@import '../utilities/themes';
 
 .App {
     text-align: center;
