+++ src/pages/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Layout, Row, Col } from 'antd';
 
-import './Home.scss';
+import '../css/main.scss';
 
 import homeImg from '../images/home.jpg';
