+++ src/pages/PageNotFound.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Layout, Row, Col } from 'antd';
-import './PageNotFound.scss';
+import '../css/main.scss';
 
 import looking from '../images/bitmoji/looking.png';
